Move connection close logic out of for loop
I'm currently experiencing a memory leak for the nebula monitoring task. While the "ClosePeer" method was only called if the connection was successful it seems to me that it also needs to be closed in case of an error. This is just a guess though and needs to be tested.